refactor: 캔버스 서비스에 워커 스레드 적용
